@@ -1,5 +1,4 @@
-﻿using Microsoft.AspNetCore.Http.Authentication;
-using Microsoft.AspNetCore.Identity;
+﻿using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
@@ -57,11 +56,11 @@ namespace KyoS.Web.Models
         public string ConfirmPassword { get; set; }        
     }
 
-    public class ManageLoginsViewModel
+    /*public class ManageLoginsViewModel
     {
         public IList<UserLoginInfo> CurrentLogins { get; set; }
         public IList<AuthenticationDescription> OtherLogins { get; set; }
-    }
+    }*/
 
     public class FactorViewModel
     {
