@@ -111,7 +111,7 @@
     <script type="text/javascript">
                         $(document).ready(function () {
                             $('#MyTable').DataTable();
-
+                            
                             // Delete item
                             var item_to_delete;
 
